// mrMr/projects/mr-mr2/src/lib/models/error-log.ts
export class ErrorLog {
  User: string;
  Token: string;
  Server: string;
  Time: Date;
  Name: string;
  Message: string;
  Details: string;
  Status?: number;
  StatusText?: string;
  Url?: string;
  Body?: any;
  Method?: string;
  Page: string;
  WebBrowser: string;
  Culture: string;
  CookiesEnabled: boolean;
  MaxTouchPoints: number;
  Platform: string;
}